Drop unused views and tests
