+++ care/facility/api/serializers/consultation_diagnosis.py
@@ -7,7 +7,6 @@ from care.facility.models import (
     ConsultationDiagnosis,
 )
 from care.facility.models.icd11_diagnosis import ICD11Diagnosis
-from care.facility.static_data.icd11 import get_icd11_diagnosis_object_by_id
 from care.users.api.serializers.user import UserBaseMinimumSerializer
 
 
@@ -32,7 +31,7 @@ class ConsultationDiagnosisSerializer(serializers.ModelSerializer):
     created_by = UserBaseMinimumSerializer(read_only=True)
 
     def get_diagnosis_object(self, obj):
-        return get_icd11_diagnosis_object_by_id(obj.diagnosis_id, as_dict=True)
+        return None
 
     class Meta:
         model = ConsultationDiagnosis
