+++ care/facility/models/patient.py
@@ -37,7 +37,6 @@ from care.facility.models.patient_base import (
     REVERSE_ROUTE_TO_FACILITY_CHOICES,
 )
 from care.facility.models.patient_consultation import PatientConsultation
-from care.facility.static_data.icd11 import get_icd11_diagnoses_objects_by_ids
 from care.users.models import GENDER_CHOICES, REVERSE_GENDER_CHOICES, User
 from care.utils.models.base import BaseManager, BaseModel
 from care.utils.models.validators import mobile_or_landline_number_validator
@@ -592,7 +591,7 @@ class PatientRegistration(PatientBaseModel, PatientPermissionMixin):
         return self.strftime("%H:%M")
 
     def format_diagnoses(self):
-        diagnoses = get_icd11_diagnoses_objects_by_ids(self)
+        diagnoses = []
         return ", ".join([diagnosis["label"] for diagnosis in diagnoses])
 
     CSV_MAKE_PRETTY = {
