+++ care/facility/tests/test_patient_consultation_api.py
@@ -9,10 +9,6 @@ from care.facility.api.serializers.patient_consultation import MIN_ENCOUNTER_DAT
 from care.facility.models.bed import Bed
 from care.facility.models.encounter_symptom import Symptom
 from care.facility.models.file_upload import FileUpload
-from care.facility.models.icd11_diagnosis import (
-    ConditionVerificationStatus,
-    ICD11Diagnosis,
-)
 from care.facility.models.patient_base import NewDischargeReasonEnum, SuggestionChoices
 from care.facility.models.patient_consultation import (
     CATEGORY_CHOICES,
@@ -46,13 +42,6 @@ class TestPatientConsultation(TestUtils, APITestCase):
             "treatment_plan": "treatment_plan",
             "suggestion": SuggestionChoices.HI,
             "treating_physician": self.doctor.id,
-            "create_diagnoses": [
-                {
-                    "diagnosis": ICD11Diagnosis.objects.first().id,
-                    "is_principal": False,
-                    "verification_status": ConditionVerificationStatus.CONFIRMED,
-                }
-            ],
             "create_symptoms": [
                 {
                     "symptom": Symptom.FEVER,
@@ -443,142 +432,6 @@ class TestPatientConsultation(TestUtils, APITestCase):
 
         res = self.update_consultation(consultation, category="MILD", suggestion="A")
         self.assertEqual(res.status_code, status.HTTP_400_BAD_REQUEST)
-
-    def test_add_diagnoses_and_duplicate_diagnoses(self):
-        consultation = self.create_admission_consultation(
-            suggestion=SuggestionChoices.A,
-            encounter_date=make_aware(datetime.datetime(2020, 4, 1, 15, 30, 00)),
-        )
-        diagnosis = ICD11Diagnosis.objects.all()[0].id
-        res = self.add_diagnosis(
-            consultation,
-            diagnosis=diagnosis,
-            is_principal=True,
-            verification_status=ConditionVerificationStatus.CONFIRMED,
-        )
-        self.assertEqual(res.status_code, status.HTTP_201_CREATED)
-        res = self.add_diagnosis(
-            consultation,
-            diagnosis=diagnosis,
-            is_principal=True,
-            verification_status=ConditionVerificationStatus.PROVISIONAL,
-        )
-        self.assertEqual(res.status_code, status.HTTP_400_BAD_REQUEST)
-
-    def test_add_diagnosis_inactive(self):
-        consultation = self.create_admission_consultation(
-            suggestion=SuggestionChoices.A,
-            encounter_date=make_aware(datetime.datetime(2020, 4, 1, 15, 30, 00)),
-        )
-        diagnosis = ICD11Diagnosis.objects.first().id
-        res = self.add_diagnosis(
-            consultation,
-            diagnosis=diagnosis,
-            is_principal=False,
-            verification_status=ConditionVerificationStatus.ENTERED_IN_ERROR,
-        )
-        self.assertEqual(res.status_code, status.HTTP_400_BAD_REQUEST)
-        res = self.add_diagnosis(
-            consultation,
-            diagnosis=diagnosis,
-            is_principal=True,
-            verification_status=ConditionVerificationStatus.REFUTED,
-        )
-        self.assertEqual(res.status_code, status.HTTP_400_BAD_REQUEST)
-
-    def mark_inactive_diagnosis_as_principal(self):
-        consultation = self.create_admission_consultation(
-            suggestion=SuggestionChoices.A,
-            encounter_date=make_aware(datetime.datetime(2020, 4, 1, 15, 30, 00)),
-        )
-        diagnosis = ICD11Diagnosis.objects.first().id
-        res = self.add_diagnosis(
-            consultation,
-            diagnosis=diagnosis,
-            is_principal=False,
-            verification_status=ConditionVerificationStatus.CONFIRMED,
-        )
-        self.assertEqual(res.status_code, status.HTTP_201_CREATED)
-        res = self.edit_diagnosis(
-            consultation,
-            res.data["id"],
-            verification_status=ConditionVerificationStatus.REFUTED,
-        )
-        self.assertEqual(res.status_code, status.HTTP_200_OK)
-        res = self.edit_diagnosis(
-            consultation,
-            res.data["id"],
-            is_principal=True,
-        )
-        self.assertEqual(res.status_code, status.HTTP_400_BAD_REQUEST)
-
-    def test_change_diagnosis(self):
-        consultation = self.create_admission_consultation(
-            suggestion=SuggestionChoices.A,
-            encounter_date=make_aware(datetime.datetime(2020, 4, 1, 15, 30, 00)),
-        )
-        res = self.add_diagnosis(
-            consultation,
-            diagnosis=ICD11Diagnosis.objects.all()[0].id,
-            is_principal=False,
-            verification_status=ConditionVerificationStatus.CONFIRMED,
-        )
-        self.assertEqual(res.status_code, status.HTTP_201_CREATED)
-        res = self.edit_diagnosis(
-            consultation,
-            res.data["id"],
-            diagnosis=ICD11Diagnosis.objects.all()[1].id,
-            is_principal=True,
-            verification_status=ConditionVerificationStatus.PROVISIONAL,
-        )
-        self.assertEqual(res.status_code, status.HTTP_400_BAD_REQUEST)
-
-    def test_add_multiple_principal_diagnosis(self):
-        consultation = self.create_admission_consultation(
-            suggestion=SuggestionChoices.A,
-            encounter_date=make_aware(datetime.datetime(2020, 4, 1, 15, 30, 00)),
-        )
-        res = self.add_diagnosis(
-            consultation,
-            diagnosis=ICD11Diagnosis.objects.all()[0].id,
-            is_principal=True,
-            verification_status=ConditionVerificationStatus.CONFIRMED,
-        )
-        self.assertEqual(res.status_code, status.HTTP_201_CREATED)
-        res = self.add_diagnosis(
-            consultation,
-            diagnosis=ICD11Diagnosis.objects.all()[1].id,
-            is_principal=True,
-            verification_status=ConditionVerificationStatus.PROVISIONAL,
-        )
-        self.assertEqual(res.status_code, status.HTTP_400_BAD_REQUEST)
-
-    def test_add_principal_edit_as_inactive_add_principal(self):
-        consultation = self.create_admission_consultation(
-            suggestion=SuggestionChoices.A,
-            encounter_date=make_aware(datetime.datetime(2020, 4, 1, 15, 30, 00)),
-        )
-        res = self.add_diagnosis(
-            consultation,
-            diagnosis=ICD11Diagnosis.objects.all()[0].id,
-            is_principal=True,
-            verification_status=ConditionVerificationStatus.CONFIRMED,
-        )
-        self.assertEqual(res.status_code, status.HTTP_201_CREATED)
-        res = self.edit_diagnosis(
-            consultation,
-            res.data["id"],
-            verification_status=ConditionVerificationStatus.ENTERED_IN_ERROR,
-        )
-        self.assertEqual(res.status_code, status.HTTP_200_OK)
-        self.assertFalse(res.data["is_principal"])
-        res = self.add_diagnosis(
-            consultation,
-            diagnosis=ICD11Diagnosis.objects.all()[1].id,
-            is_principal=True,
-            verification_status=ConditionVerificationStatus.PROVISIONAL,
-        )
-        self.assertEqual(res.status_code, status.HTTP_201_CREATED)
 
     def test_create_consultations_with_same_patient_no_in_different_facilities(self):
         facility2 = self.create_facility(
