+++ docs/django-commands/configuration.rst
@@ -18,10 +18,6 @@ All Django commands must be executed by invoking :code:`python manage.py`, so fo
 |                     |                                                               |                                                                                                                                                                                                                                                                                                                                  |
 |                     |                                                               | Example Invocation: :code:`python manage.py load_data kerala`                                                                                                                                                                                                                                                                    |
 +---------------------+---------------------------------------------------------------+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
-| load_medicines_data | None                                                          | This command is used to load medibase medicines into the database from medibase.json                                                                                                                                                                                                                                             |
-|                     |                                                               |                                                                                                                                                                                                                                                                                                                                  |
-|                     |                                                               | Example Invocation: :code:`python manage.py load_medicines_data`                                                                                                                                                                                                                                                                 |
-+---------------------+---------------------------------------------------------------+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
 | seed_data           | None                                                          | This command loads all static data that the application needs to run, this includes inventory seed data for Oxygen, Medical Tests data, etc. All this data can be changed by the superuser from the admin panel.                                                                                                                 |
 |                     |                                                               |                                                                                                                                                                                                                                                                                                                                  |
 |                     |                                                               | Example Invocation: :code:`python manage.py seed_data`                                                                                                                                                                                                                                                                           |
